k8s 管理 接口代理功能

+++ xiaochen-modules/xiaochen-k8s-manage/src/main/java/com/zvz09/xiaochen/k8s/manage/service/impl/IClusterServiceImpl.java
@@ -5,6 +5,7 @@ import com.zvz09.xiaochen.common.core.exception.BusinessException;
 import com.zvz09.xiaochen.common.web.util.StringUtils;
 import com.zvz09.xiaochen.k8s.manage.domain.dto.ClusterDTO;
 import com.zvz09.xiaochen.k8s.manage.domain.entity.Cluster;
+import com.zvz09.xiaochen.k8s.manage.filter.ProxyRoutes;
 import com.zvz09.xiaochen.k8s.manage.service.IClusterService;
 import com.zvz09.xiaochen.k8s.manage.mapper.ClusterMapper;
 import com.zvz09.xiaochen.k8s.manage.service.K8sService;
@@ -21,6 +22,7 @@ import org.springframework.stereotype.Service;
 public class IClusterServiceImpl extends ServiceImpl<ClusterMapper, Cluster> implements IClusterService {
 
     private final K8sService k8sService;
+    private final ProxyRoutes proxyRoutes;
 
     @Override
     public void create(ClusterDTO clusterDTO) {
@@ -36,12 +38,18 @@ public class IClusterServiceImpl extends ServiceImpl<ClusterMapper, Cluster> imp
         cluster.setApiServer(clusterDTO.getApiServer());
         cluster.setToken(clusterDTO.getToken());
 
+        proxyRoutes.addRoute(cluster);
         this.save(cluster);
     }
 
     @Override
     public void delete(Long id) {
-        this.baseMapper.deleteById(id);
+        Cluster cluster = this.getById(id);
+        if(cluster != null){
+            proxyRoutes.removeRoute(cluster.getApiServer());
+            this.baseMapper.deleteById(id);
+        }
+
     }
 }
